refactor(admin): move Studio ownership to canonical routes
