--- apps/admin/app/admin/studio/agents/page.tsx
@@ -1,9 +1,0 @@
-import { Metadata } from 'next';
-import AgentsClient from './AgentsClient';
-
-export const dynamic = 'force-dynamic';
-export const metadata: Metadata = { title: 'AI Agents | Dev Studio' };
-
-export default function AgentsPage() {
-  return <AgentsClient />;
-}
