--- apps/admin/app/admin/studio/media/layout.tsx
@@ -1,18 +1,0 @@
-import { Metadata } from 'next';
-import { PLATFORM_DEFAULTS } from '@/lib/config/platform-config';
-
-export const metadata: Metadata = {
-  title: `Admin | Media studio | ${PLATFORM_DEFAULTS.orgName}`,
-  description: `${PLATFORM_DEFAULTS.orgName} - Career training and workforce development programs.`,
-  alternates: {
-    canonical: 'https://www.elevateforhumanity.org/admin/media-studio',
-  },
-  robots: {
-    index: false,
-    follow: false,
-  },
-};
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
-}
