--- apps/admin/app/admin/studio/media/page.tsx
@@ -1,15 +1,0 @@
-// Server component — enforces admin auth at the route boundary.
-// The client component never renders for unauthorized users.
-import AdminClientPage from '@/components/admin/AdminClientPage';
-import MediaStudioClient from './MediaStudioClient';
-
-export const dynamic = 'force-dynamic';
-export const revalidate = 60;
-
-export default function Page() {
-  return (
-    <AdminClientPage>
-      <MediaStudioClient />
-    </AdminClientPage>
-  );
-}
