+++ apps/admin/app/studio/agents/page.tsx
@@ -1,1 +1,9 @@
-export { default, dynamic, metadata } from '../../admin/studio/agents/page';
+import { Metadata } from 'next';
+import AgentsClient from './AgentsClient';
+
+export const dynamic = 'force-dynamic';
+export const metadata: Metadata = { title: 'AI Agents | Dev Studio' };
+
+export default function AgentsPage() {
+  return <AgentsClient />;
+}
