+++ apps/admin/app/studio/builds/page.tsx
@@ -1,1 +1,9 @@
-export { default, dynamic, metadata } from '../../admin/studio/builds/page';
+import { Metadata } from 'next';
+import BuildsClient from './BuildsClient';
+
+export const dynamic = 'force-dynamic';
+export const metadata: Metadata = { title: 'Builds | Dev Studio' };
+
+export default function BuildsPage() {
+  return <BuildsClient />;
+}
