+++ apps/admin/app/studio/deployments/page.tsx
@@ -1,1 +1,9 @@
-export { default, dynamic, metadata } from '../../admin/studio/deployments/page';
+import { Metadata } from 'next';
+import DeploymentsClient from './DeploymentsClient';
+
+export const dynamic = 'force-dynamic';
+export const metadata: Metadata = { title: 'Deployments | Dev Studio' };
+
+export default function DeploymentsPage() {
+  return <DeploymentsClient />;
+}
