+++ apps/admin/app/studio/memory/page.tsx
@@ -1,1 +1,9 @@
-export { default, dynamic, metadata } from '../../admin/studio/memory/page';
+import { Metadata } from 'next';
+import MemoryClient from './MemoryClient';
+
+export const dynamic = 'force-dynamic';
+export const metadata: Metadata = { title: 'AI Memory | Dev Studio' };
+
+export default function MemoryPage() {
+  return <MemoryClient />;
+}
