+++ apps/admin/app/studio/settings/page.tsx
@@ -1,1 +1,9 @@
-export { default, dynamic, metadata } from '../../admin/studio/settings/page';
+import { Metadata } from 'next';
+import SettingsClient from './SettingsClient';
+
+export const dynamic = 'force-dynamic';
+export const metadata: Metadata = { title: 'Settings | Dev Studio' };
+
+export default function SettingsPage() {
+  return <SettingsClient />;
+}
