+++ apps/admin/app/studio/tasks/page.tsx
@@ -1,1 +1,9 @@
-export { default, dynamic, metadata } from '../../admin/studio/tasks/page';
+import { Metadata } from 'next';
+import TasksClient from './TasksClient';
+
+export const dynamic = 'force-dynamic';
+export const metadata: Metadata = { title: 'AI Tasks | Dev Studio' };
+
+export default function TasksPage() {
+  return <TasksClient />;
+}
