+++ apps/admin/app/studio/workflows/[id]/WorkflowDetailClient.tsx
